move all accessibility edits into their own file
